+++ lib/action_policy/behaviours/memoized.rb
@@ -19,9 +19,6 @@ module ActionPolicy
     #
     #   policy.equal?(policy_for(record, with: CustomPolicy)) #=> false
     module Memoized
-      require "action_policy/ext/policy_cache_key"
-      using ActionPolicy::Ext::PolicyCacheKey
-
       class << self
         def prepended(base)
           base.prepend InstanceMethods
@@ -32,13 +29,12 @@ module ActionPolicy
 
       module InstanceMethods # :nodoc:
         def policy_for(record:, **opts)
-          __policy_memoize__(record, opts) { super(record: record, **opts) }
+          __policy_memoize__(record, **opts) { super(record: record, **opts) }
         end
       end
 
-      def __policy_memoize__(record, with: nil, namespace: nil, **_opts)
-        record_key = record._policy_cache_key(use_object_id: true)
-        cache_key = "#{namespace}/#{with}/#{record_key}"
+      def __policy_memoize__(record, **options)
+        cache_key = policy_for_cache_key(record: record, **options)
 
         return __policies_cache__[cache_key] if
           __policies_cache__.key?(cache_key)