+++ lib/action_policy/behaviours/scoping.rb
@@ -11,6 +11,8 @@ module ActionPolicy
       #   - secondly, try to infer policy class from `target` (non-raising lookup)
       #   - use `implicit_authorization_target` if none of the above works.
       def authorized_scope(target, type: nil, as: :default, scope_options: nil, **options)
+        options[:context] && (options[:context] = authorization_context.merge(options[:context]))
+
         policy = policy_for(record: target, allow_nil: true, **options)
         policy ||= policy_for(record: implicit_authorization_target!, **options)